refactor: domain, view package 분리
