+++ src/main/java/racingcar/view/OutputView.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import racingcar.domain.CarDto;
 
 public class OutputView {
 	private static final String GAME_RESULT_TITLE = "\n실행 결과";
