+++ src/test/java/racingcar/RefereeTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.Car;
+import racingcar.domain.Referee;
+
 public class RefereeTest {
 	@Test
 	public void 차_전진_판단() {
